Initial work to make the linenoise library usable on RISC OS.
The initial work here has been to create a simple library which can
be used to create a line editor which can be used in a rudimentary
way on RISC OS. We use the RISC OS VDU sequences, and update the
cursor controls so that we have behaviour which is closer to our
expectations. The hints don't quite work right because they have
no highlighting available, but I wouldn't expect them to be used
in much. The multiline editing has not been updated at all.

The RISC OS cursors are handled by a small library which manipulates
the cursor input state. This is currently using the standard mechanism
which does not prefix by 0 bytes. This isn't helpful as it will break
some input. However, for now it works ok.

It won't work with UTF-8 input, but that isn't a general configuration
used on RISC OS, so probably isn't a concern right now - it can be
investigated later.