Refactor the linenoise file to allow multiple configurations.
The line editing configuration is now able to be used with multiple
configurations simultaneously. This should allow the interface to be
used by multiple concurrent systems - obviously with different
configuration for their state. There may need to be some more locking
in the code to be truely reentrant, but that may not be a necessary
requirement.