Add the option to not rewrite the prompt for command lines.
Instead of jumping to the start of the line, we now write the editing
by moving backwards on the line, to make it easier to edit within the
RISC OS environment where prompts are output independant of the line
editing system.

This is still a little wrong because it doesn't take account of the
offset on the line when working out the start of the output for
clearing the lines. This can probably be worked out, but hasn't been
done yet.