Add a RISC OS module that claims the ReadLineV.
We now have a RISC OS module that traps ReadLineV and allows the
system to read lines using the LineNoise library. It might not be
perfect - it almost certainly isn't perfect - but it does give us
a simple implementation that we can use for editing stuff.

It certainly needs more work, but at least it tries to do the
right thing!