Fix for escape return in module; improved error reporting.
The Escape key return from the module's ReadLineV was broken. It
didn't return the C flag properly - returning '29' instead of 'bit 29'.
This means that the module now responds properly on RISC OS Classic
when you press escape.

The error handling in the main library has been improved so that it
includes the reason in errno for all (?) the mechanisms by which the
library can exit. This should make it easier to see why the call
failed.