Fix for cursor positioning using the initial cursor position.
The initial cursor position is now used, ignoring the prompt, if a
NULL prompt was given. This matches with the RISC OS behaviour.

The position we move the cursor to at the end of the line is now
correct when we're scrolling through the line.

The column width now correctly uses the window width - 1 (to avoid
scrolling) and not window width - 2 (due to the width being the max
position - 1).