Add support for the commands to control the history.
The RISC OS module can now load and save the history, list the contents,
and erase it. These commands are backed by the library providing the
original functions for the history, plus new functions to read the
lines of the history and exposing the ability to free the history
buffer.