Update the RISC OS cursors; fixed visible state, more key reports.
The ro_cursors library has been updated to handle cursors better.
For a start the state of the cursor was wrong - its visibility and
removal were inverted. That's now fixed.

We also reset all the function and control keys so that we can
hopefully report them distinctly as escaped keys.

The state is now read on startup and restored on shutdown so
hopefully we should be nicer for the user.